slightly refactor code
