+++ preact-workerd-ssr-plugin/src/plugin.ts
@@ -1,10 +1,7 @@
 import fs from "node:fs";
 import path from "node:path";
-import { IncomingMessage } from "node:http";
-import { createMiniflareInstance } from "./miniflare.js";
 import { type ViteDevServer } from "vite";
-//@ts-ignore
-import workerdBootloader from "./workerdBootloader.js.txt";
+import { createWorkerdHandler } from "./workerdHandler";
 
 export function preactWorkerdSSR() {
   return {
@@ -14,6 +11,12 @@ export function preactWorkerdSSR() {
         const handler = createWorkerdHandler({
           entrypoint: "./entry-server.jsx",
           server,
+          frameworkRequestHandlingJs: `
+            const url = request.url;
+            const renderedString = entryPoint.render(url);
+
+            return new Response(renderedString);
+          `
         });
 
         server.middlewares.use(async (req, res, next) => {
@@ -41,10 +44,8 @@ export function preactWorkerdSSR() {
 
             template = await server.transformIndexHtml(url, template);
 
-            // dispatch request to miniflare
             const ssrResponse = await handler(req);
 
-            // preact specific
             if (ssrResponse.status !== 200) {
               res.statusCode = ssrResponse.status;
               res.statusMessage = ssrResponse.statusText;
@@ -68,71 +69,5 @@ export function preactWorkerdSSR() {
         });
       };
     },
-  };
-}
-
-function createWorkerdHandler({
-  entrypoint,
-  server,
-}: {
-  entrypoint: string;
-  server: ViteDevServer;
-}) {
-  console.log("create handler");
-
-  // TODO: figure out how to get hold of Vite's server address
-  // server.httpServer.address() is null, likely because the server hasn't started yet
-  //console.log('address', config.server.httpServer.address())
-  const address = "http://localhost:5173";
-
-  const bootloader = workerdBootloader
-    .replace(/VITE_SERVER_ADDRESS/, address)
-    .replace(/WORKERD_APP_ENTRYPOINT/, entrypoint)
-    .replace(
-      /GENERATE_RESPONSE/,
-      `
-            // preact specific bits
-            const url = request.url;
-            const renderedString = entryPoint.render(url);
-
-            return new Response(renderedString);
-            `
-    );
-
-  // create miniflare instance
-  // load it with module loader code and import to the entry point
-
-  const mf = createMiniflareInstance({ script: bootloader });
-
-  // this could be in the future replaced with websockets
-  server.middlewares.use(async (request, resp, next) => {
-    if (!request.url.startsWith("/__workerd_loader/")) {
-      next();
-      return;
-    }
-
-    // the request is for the workerd loader so we need to handle it
-    const url = new URL(`http://localhost${request.url}`);
-    const moduleId = url.searchParams.get("moduleId");
-
-    const moduleCode = (await server.transformRequest(moduleId, {
-      ssr: true
-    })).code;
-    resp.writeHead(200, { "Content-Type": "text/plain" });
-    resp.end(moduleCode);
-  });
-
-  return async function workerdRequestHandler(request: IncomingMessage) {
-    let url = request.url.startsWith("/")
-      ? `http://localhost${request.url}`
-      : request.url;
-
-    // TODO: we should support POST requests with body as well
-    //       for that we need something along the lines of
-    //       https://github.com/sveltejs/kit/blob/master/packages/kit/src/exports/node/index.js#L8
-    return mf.dispatchFetch(url, {
-      headers: request.headers,
-      method: request.method || "GET",
-    });
   };
 }
