+++ preact-workerd-ssr-plugin/src/workerdHandler/workerdBootloader.js.txt
@@ -12,7 +12,6 @@ export default {
     }
 }
 
-
 const __ourPrivateModuleRegistry__ = new Map();
 
 // create our own implementation
